+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial Version
 
 - ics-dm-cli can inject wifi config into wic image file
+- ics-dm-cli can inject enrollment config into wic image file
 - ics-dm-cli can inject iotedge gateway configuration into wic image file
 - ics-dm-cli can inject iotedge leaf configuration into wic image file